test(cloudflare/containers): cover the sid collision with a deployed worker
Replace the compile-only binding-collection test with real coverage in
Container.test.ts's dev/live matrix: the async fixture (whose Container
logical id equals its env key) is deployed and driven over HTTP.

- /binding asserts env.ECHO is a durable_object_namespace, not the
  Container declaration uploaded as json
- /hello asserts the request round-trips Worker -> DO -> container

Without the fix both fail: dev dies with `Durable Object namespace
AsyncEchoObject not found`, live reports
`{"_id":"Effect","op":"alchemy/EffectClass"}`.

AsyncContainer.test.ts is folded in — Container.test.ts now covers the
same stack in both dev and live.
