+++ packages/alchemy/test/Cloudflare/Container/Container.test.ts
@@ -6,6 +6,7 @@ import { MinimumLogLevel } from "effect/References";
 import * as Schedule from "effect/Schedule";
 import * as HttpClient from "effect/unstable/http/HttpClient";
 import * as HttpClientRequest from "effect/unstable/http/HttpClientRequest";
+import AsyncContainerStack from "./fixtures/async/stack.ts";
 import EffectfulStack from "./fixtures/effectful/stack.ts";
 import ExternalStack from "./fixtures/external/stack.ts";
 import RemoteStack from "./fixtures/remote/stack.ts";
@@ -162,6 +163,56 @@ describe.concurrent.each([
       Effect.gen(function* () {
         const { url } = yield* stack;
 
+        const hello = yield* fetchReady(new URL("/hello", url), "method");
+        expect(hello).toContain("method");
+      }).pipe(logLevel),
+      { timeout },
+    );
+  });
+  /**
+   * Container bound directly on a plain async Worker's `env` (issue #953),
+   * with the Container's logical id EQUAL to the env key
+   * (`ECHO: Container("ECHO", …)`). The Worker's two halves — the
+   * `durable_object_namespace` binding and the `containers` script metadata —
+   * describe the same env entry, so they are contributed under one `sid`.
+   * Contributing them under two (the env key and the ContainerApplication's
+   * logical id) made them collide for exactly this shape, and binding rows
+   * are collapsed by sid — last write wins — so the namespace binding was
+   * silently dropped. Live, the Worker then uploads the Container declaration
+   * itself as a `json` binding (`{"_id":"Effect","op":"alchemy/EffectClass"}`)
+   * and `getContainer` dies with `t.idFromName is not a function`; in dev the
+   * local runtime holds container metadata for a class it has no namespace
+   * for and the deploy dies with
+   * `Durable Object namespace AsyncEchoObject not found`.
+   */
+  describe("async container bound on env", () => {
+    const { test, beforeAll, afterAll, deploy, destroy } = make();
+    const stack = beforeAll(deploy(AsyncContainerStack), {
+      timeout: HOOK_TIMEOUT,
+    });
+    afterAll.skipIf(!!process.env.NO_DESTROY)(destroy(AsyncContainerStack), {
+      timeout: HOOK_TIMEOUT,
+    });
+
+    test(
+      "binds the container class as a durable object namespace",
+      Effect.gen(function* () {
+        const { url } = yield* stack;
+
+        const body = yield* fetchReady(new URL("/binding", url), "kind");
+        expect(JSON.parse(body)).toEqual({ kind: "durable_object_namespace" });
+      }).pipe(logLevel),
+      { timeout },
+    );
+
+    test(
+      "serves through the container-backed DO class",
+      Effect.gen(function* () {
+        const { url } = yield* stack;
+
+        // The echo image reflects the request as JSON ("method" only appears
+        // in a real echo response, never in an error page) — proof the
+        // request went Worker → DO class → container port 8080 and back.
         const hello = yield* fetchReady(new URL("/hello", url), "method");
         expect(hello).toContain("method");
       }).pipe(logLevel),
